Safari-611.1.10/LayoutTests/imported/w3c/web-platform-tests/IndexedDB/idbcursor_update_index8-expected.txt


PASS IDBCursor.update() - index - throw InvalidStateError when the cursor is being iterated
